Assignment-2/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
certifi==2025.4.26
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
distro==1.9.0
griffe==1.7.3
h11==0.16.0
-e file:///home/Saeed/Desktop/practice/python-crash/Q-2-Assignment-2/hello-agent
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.0
idna==3.10
jiter==0.9.0
mcp==1.6.0
openai==1.76.2
openai-agents==0.0.14
pydantic==2.11.4
pydantic-core==2.33.2
pydantic-settings==2.9.1
python-dotenv==1.1.0
requests==2.32.3
sniffio==1.3.1
sse-starlette==2.3.3
starlette==0.46.2
tqdm==4.67.1
types-requests==2.32.0.20250328
typing-extensions==4.13.2
typing-inspection==0.4.0
urllib3==2.4.0
uvicorn==0.34.2
